requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-phonenumber-field==6.1.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
phonenumbers==8.12.47
psycopg2-binary==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
